requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.2.0
langchain-community==0.2.0
langchain-chroma
chromadb==0.5.0
python-dotenv
pypdf
pydantic